fix: add ButtonError component
