+++ src/App.tsx
@@ -1,25 +1,17 @@
 import './App.css';
-import Button from './components/Button';
+import ButtonError from './components/ButtonError';
 import ErrorBoundary from './components/ErrorBoundary';
 import Results from './components/Results';
 import TopControls from './components/TopControls';
 import { SearchProvider } from './context/SearchContext';
 
 function App() {
-  /**
-   * clickHandler
-   */
-
-  // const clickHandler = (): void => {
-  //   throw new Error('Something went wrong!');
-  // };
-
   return (
     <SearchProvider>
       <TopControls />
       <ErrorBoundary>
         <Results />
-        <Button type="button" name="Error" />
+        <ButtonError type="button" name="Error" />
       </ErrorBoundary>
     </SearchProvider>
   );
