+++ src/components/Button.tsx
@@ -1,37 +1,12 @@
 import React from 'react';
 
-interface Props {
+interface PropsI {
   name: string;
-  type: string;
-  onClick?: () => void;
-  // disabled?: boolean;
-  // className?: string;
-  // children?: React.ReactNode;
-  // style?: React.CSSProperties;
-  // [key: string]: any; // for additional props
 }
 
-interface State {
-  hasError: boolean;
-}
-
-class Button extends React.Component<Props, State> {
-  constructor(props: Props) {
-    super(props);
-    this.state = {
-      hasError: false,
-    };
-    this.handleClick = this.handleClick.bind(this);
-  }
-
-  handleClick = (): void => {
-    this.setState({ hasError: true });
-  };
+class Button extends React.Component<PropsI> {
   render() {
-    if (this.state.hasError) {
-      throw new Error('Expected error');
-    }
-    return <button onClick={this.handleClick}>{this.props.name}</button>;
+    return <button>{this.props.name}</button>;
   }
 }
 export default Button;
